requirements.txt
# FastAPI y Servidor
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
requests==2.32.3
starlette==0.46.0

# Machine Learning (PyTorch)
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
numpy==1.26.4
scikit-learn==1.6.1
joblib==1.4.2

# Procesamiento de imágenes
pillow==11.1.0
opencv-python==4.11.0.86

# Manejo de Datos y Visualización
pandas==2.2.3
matplotlib==3.10.0
seaborn==0.13.2

# TensorFlow 
tensorflow==2.16.2  # Para Windows/Linux

# Utilidades y otras dependencias
protobuf==4.25.6
packaging==24.2
termcolor==2.5.0
tqdm==4.67.1